pyproject.toml
[project]
name = "py-auth"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "rumia", email = "github@rumia-ch.uk" }
]
dependencies = [
    "cryptography>=45.0.4",
    "fido2>=2.0.0",
    "python-dotenv>=1.1.1",
    "werkzeug>=3.1.3",
    "qrcode>=8.2",
    "cbor2>=5.6.5",
    "flask-wtf>=1.2.2",
    "flask-sqlalchemy>=3.1.1",
    "flask-migrate>=4.1.0",
    "pyotp>=2.9.0",
    "pillow>=11.2.1",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/py_auth"]
